+++ .eslintrc
@@ -1,8 +1,12 @@
 {
-    "extends": "airbnb-base",
+    "extends": "airbnb",
+    "env": {
+        "node": true,
+        "es6": true,
+        "browser": true
+    },
     "rules": {
-        "linebreak-style":0,
-        "global-require": 0,
-        "eslint linebreak-style": [0, "error", "windows"]
+        "Linebreak-style": 0,
+        "eslint comma-dangle": [2, "always-multiline"]
     }
-}
+ }
