+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon app.js",
-    "test": "nyc --reporter=html --reporter=text mocha --require babel-register --timeout 10000 --exit  ./tests/test",
+    "start": "nodemon server/app.js",
+    "test": "nyc --reporter=html --reporter=text mocha --require babel-register --timeout 10000 --exit  server/tests/test",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
@@ -33,7 +33,8 @@
     "jsonwebtoken": "^8.5.1",
     "mocha": "^6.2.0",
     "nodemon": "^1.19.1",
-    "nyc": "^14.1.1"
+    "nyc": "^14.1.1",
+    "swagger-ui-express": "^4.0.7"
   },
   "devDependencies": {
     "eslint": "^5.16.0",
