+++ server/app.js
@@ -1,2 +1,2 @@
 require('babel-register');
-require('./server');
+require('./serve');
