+++ server/serve.js
@@ -1,8 +1,11 @@
 import express from 'express';
-import route from './src/router/route';
+import route from './router/route';
+import swaggerUi from 'swagger-ui-express';
+import swaggerDoc from '../swagger.json';
 
 const app = express();
 app.use(express.json());
+app.use('/freeMentor', swaggerUi.serve, swaggerUi.setup(swaggerDoc));
 const port = process.env.PORT || 3000;
 app.use(route);
